Update to change obs dtype

+++ apop/envs/classic_control/acrobot.py
@@ -80,4 +80,4 @@ class ContinuousAcrobotEnv(AcrobotEnv):
 
         if self.render_mode == "human":
             self.render()
-        return (np.array(self.state), reward, terminated, False, {})
+        return (np.array(self.state, dtype=np.float32), reward, terminated, False, {})
